add @AutoClose
This annotation marks _auto-closeable_ beans. If the class of an instance
of auto-closeable bean implements `AutoCloseable`, the `close()` method
is called during bean destruction (after `@PreDestroy` callbacks or
disposer method invocation, but before destroying dependent instances).

The invocation of `close()` during bean destruction is not a business method
invocation and so is not intercepted or decorated.